Move sequence selector into same module as main window
